requirements.txt
absl-py==2.1.0
ale-py==0.11.2
altair==5.5.0
amqp==5.3.1
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.11.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==26.1.0
Automat==24.8.1
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.2
cachelib==0.13.0
cachetools==5.5.1
celery==5.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromedriver-binary==135.0.7022.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
constantly==23.10.4
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.12
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-mongodb-engine==0.6.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djangotoolbox==1.8.0
djongo==1.2.31
dnspython==2.7.0
drf-spectacular==0.29.0
drf-yasg==1.21.15
et_xmlfile==2.0.0
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
folium==0.20.0
fonttools==4.56.0
fpdf==1.7.2
fqdn==1.5.1
fsspec==2024.12.0
fuzzywuzzy==0.18.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.24.1
google-api-python-client==2.48.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
Google-Images-Search==1.4.7
google-pasta==0.2.0
google_images_download==2.8.0
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.68.1
gunicorn==25.3.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
ImageHash==4.3.2
imbalanced-learn==0.13.0
incremental==24.7.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.7.0
kiwisolver==1.4.8
kombu==5.6.2
Levenshtein==0.27.1
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
MiniSom==2.3.5
mistune==3.1.1
ml-dtypes==0.4.1
mlxtend==0.23.4
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.19.0
mysqlclient==2.2.8
namex==0.0.8
narwhals==2.6.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
npm==0.1.1
numpy==2.0.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.13.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==26.0
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
Protego==0.4.0
proto-plus==1.26.0
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7
pyfiglet==0.8.post1
pygame==2.6.1
Pygments==2.18.0
pyjokes==0.8.3
PyJWT==2.12.1
pymongo==4.10.1
pynput==1.8.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypiwin32==223
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-Levenshtein==0.27.1
python-resize-image==1.1.20
pyttsx3==2.98
pytz==2025.1
PyWavelets==1.8.0
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.3
pyzmq==26.2.1
Quandl==3.7.0
queuelib==1.7.0
RapidFuzz==3.14.1
redis==7.4.0
referencing==0.37.0
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.30.0
rsa==4.9
scikit-fuzzy==0.5.0
scikit-learn==1.6.0
scipy==1.14.1
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
service-identity==24.2.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.5
stable_baselines3==2.7.0
stack-data==0.6.3
streamlit==1.50.0
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tldextract==5.1.3
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.1
Twisted==24.11.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
virtualenv==20.28.1
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.12.0
wrapt==1.17.0
wsproto==1.2.0
xgboost==2.1.4
xyzservices==2025.11.0
zope.interface==7.2
